backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
solders>=0.21.0
PyNaCl>=1.5.0
pydantic>=2.5.0
python-jose>=3.3.0
httpx>=0.26.0
base58>=2.1.0
